moved user info and login/logout to toolbar
