+++ client/src/Main.jsx
@@ -31,8 +31,8 @@ const container = document.getElementById('root');
 const root = createRoot(container);
 root.render(
   <BrowserRouter>
-    <NavLinks/>
     <UserAuthContextProvider>
+      <NavLinks/>
       <Routes>
         <Route path='/home' element={<Homepage />}/>
         <Route path='/login' element={<Login />} />
